fix: refactor; fix testcase
